+++ tests/de_test.go
@@ -5,7 +5,6 @@ import (
 	"time"
 
 	"github.com/SerhiiCho/timeago/v3"
-	"github.com/SerhiiCho/timeago/v3/config"
 )
 
 const langDe = "de"
@@ -62,7 +61,7 @@ func TestParseDe(t *testing.T) {
 
 	for _, tc := range cases {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
-			timeago.Configure(&config.Config{Language: langDe})
+			timeago.Configure(&timeago.Config{Language: langDe})
 
 			if res := timeago.Parse(tc.date); res != tc.result {
 				test.Errorf("Result must be %s, but got %s instead", tc.result, res)
@@ -92,7 +91,7 @@ func TestParseDeWithSeconds(t *testing.T) {
 
 	for _, tc := range cases {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
-			timeago.Configure(&config.Config{Language: langDe})
+			timeago.Configure(&timeago.Config{Language: langDe})
 
 			if res := timeago.Parse(tc.date); res != tc.result[0] && res != tc.result[1] {
 				test.Errorf("Result must be %s or %s, but got %s instead", tc.result[0], tc.result[1], res)