+++ tests/en_test.go
@@ -5,7 +5,6 @@ import (
 	"time"
 
 	"github.com/SerhiiCho/timeago/v3"
-	"github.com/SerhiiCho/timeago/v3/config"
 )
 
 const langEn = "en"
@@ -56,7 +55,7 @@ func TestParseEn(t *testing.T) {
 
 	for _, tc := range cases {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
-			timeago.Configure(&config.Config{Language: langEn})
+			timeago.Configure(&timeago.Config{Language: langEn})
 
 			if res := timeago.Parse(tc.date); res != tc.result {
 				test.Errorf("Result must be %s, but got %s instead", tc.result, res)
@@ -85,7 +84,7 @@ func TestParseEnWithSeconds(t *testing.T) {
 
 	for _, tc := range cases {
 		t.Run("result for "+tc.date.String(), func(test *testing.T) {
-			timeago.Configure(&config.Config{Language: langEn})
+			timeago.Configure(&timeago.Config{Language: langEn})
 
 			if res := timeago.Parse(tc.date); res != tc.result[0] && res != tc.result[1] {
 				test.Errorf("Result must be %s or %s, but got %s instead", tc.result[0], tc.result[1], res)